Extract ledgers update function
