+++ src/state/fundingPayment/reducer.js
@@ -1,8 +1,8 @@
 // https://docs.bitfinex.com/v2/reference#ledgers
-import _get from 'lodash/get'
 
 import queryTypes from 'state/query/constants'
 import authTypes from 'state/auth/constants'
+import { updateLedgers } from 'state/ledgers/utils'
 import {
   addSymbol,
   baseSymbolState,
@@ -12,7 +12,6 @@ import {
   setSymbols,
   setTimeRange,
 } from 'state/reducers.helper'
-import { mapSymbol, mapDescription } from 'state/symbols/utils'
 
 import types from './constants'
 
@@ -25,53 +24,8 @@ const TYPE = queryTypes.MENU_FPAYMENT
 export function fpaymentReducer(state = initialState, action) {
   const { type, payload } = action
   switch (type) {
-    case types.UPDATE_FPAYMENT: {
-      const res = _get(payload, ['data', 'res'])
-      if (!res) {
-        return {
-          ...state,
-          dataReceived: true,
-        }
-      }
-      const { existingCoins } = state
-      const updateCoins = [...existingCoins]
-      const entries = res.map((entry) => {
-        const {
-          amount,
-          amountUsd,
-          balance,
-          balanceUsd,
-          currency,
-          description,
-          id,
-          mts,
-          wallet,
-        } = entry
-        const mappedCurrency = mapSymbol(currency)
-        // save new symbol to updateCoins list
-        if (updateCoins.indexOf(mappedCurrency) === -1) {
-          updateCoins.push(mappedCurrency)
-        }
-        return {
-          id,
-          currency: mappedCurrency,
-          mts,
-          amount,
-          amountUsd,
-          balance,
-          balanceUsd,
-          description: mapDescription(description),
-          wallet,
-        }
-      })
-      return {
-        ...state,
-        dataReceived: true,
-        entries: [...state.entries, ...entries],
-        existingCoins: updateCoins.sort(),
-        pageLoading: false,
-      }
-    }
+    case types.UPDATE_FPAYMENT:
+      return updateLedgers(state, payload)
     case types.FETCH_FAIL:
       return fetchFail(state)
     case types.ADD_SYMBOL:
